Further updates, including the capability to sell shares as a function of buying.

+++ backend/handlers/bets/betutils/betutils.go
@@ -45,7 +45,7 @@ func init() {
 
 // CheckMarketStatus checks if the market is resolved or closed.
 // It returns an error if the market is not suitable for placing a bet.
-func CheckMarketStatus(db *gorm.DB, marketID int64) error {
+func CheckMarketStatus(db *gorm.DB, marketID uint) error {
 	var market models.Market
 	if result := db.First(&market, marketID); result.Error != nil {
 		if errors.Is(result.Error, gorm.ErrRecordNotFound) {
