+++ backend/models/bets.go
@@ -9,10 +9,10 @@ import (
 type Bet struct {
 	gorm.Model
 	Action   string    `json:"action"`
-	ID       int64     `json:"id" gorm:"primary_key"`
+	ID       uint      `json:"id" gorm:"primary_key"`
 	Username string    `json:"username"`
 	User     User      `gorm:"foreignKey:Username;references:Username"`
-	MarketID int64     `json:"marketId"`
+	MarketID uint      `json:"marketId"`
 	Market   Market    `gorm:"foreignKey:ID;references:MarketID"`
 	Amount   int64     `json:"amount"`
 	PlacedAt time.Time `json:"placedAt"`
